configs: prevent infinite loops in dynamicconfig generation
Summary:
There's an issue where hg foo could spawn hg dynamicconfig which could
spawn hg version (via the telemetry wrapper) which could spawn hg dynamicconfig.
If the config generation fails then this could enter an infiniteloop, since the
mtime of the file would not be updated appropriately, so we will always think we
need to regenerate.

Let's prevent this by setting an environment variable when spawning the
background process.

Reviewed By: quark-zju

Differential Revision: D21885363

fbshipit-source-id: ba4938c926d1219985813a521fec412097df4b4a